Add --preproc, and concatenation in the presence of chapters
Tis lets you say

11:1 Chapter 1 of title 11

to encode that chapter as an alias to the existing --chapter option.

This is so that you can concatenate them! You could already say

1+5+9 Three titles concatenated

but now you can also say

11:1+11:5:11:9 Three chapters concatenated

to concat three chapters. (Of course, they must all use the same
options, since there's only one space for a :: --options string; hence
the new notation.)

We also add --preproc, which is like --postproc except that the
preprocessor is invoked on the input before it is transformed by
vapoursynth. As with --postproc, and indeed concatenation and
vapoursynthing, this only works with things that aren't pulled straight
off the source medium and transcoded: i.e. in conjunction with
--makemkv or --mkvdir.